feat(host-extension): extract PluginRenderer (Phase 3e)
The content-script DomJson → DOM renderer is now in
@tlsn/host-extension/content. Mirrors the same drag/persist/click-bridge
behavior the extension always shipped; the platform-specific
postMessage decision (where clicks go) is injected by the consumer.

PluginRenderer (~190 lines, new file)
- `renderPluginUI(json, windowId, { onPluginAction, containerId? })`
  mounts the tree under `#tlsn-plugin-container` (configurable).
- `createNode(json, windowId, onPluginAction)` recursively walks
  DomJson, applies the same ALLOWED_ELEMENT_TYPES allowlist (33
  element types), handles className / id / style / inputType / checked
  / value / placeholder / disabled, wires onclick → callback,
  recursively builds children, marks draggable nodes via
  `data-tlsn-draggable`.
- `makeDraggable(el)` ports the extension's drag-from-first-child
  behavior (which lets the user move the overlay around with bottom/
  right → top/left conversion preserved across re-renders).
- `unmountPluginUI(containerId?)` for cleanup.
- The click bridge is a `PluginActionHandler` callback supplied by the
  consumer rather than a hardcoded `browser.runtime.sendMessage(…)`;
  same shape host-cli uses with its Playwright `exposeFunction` bridge.

Extension consumer (Content/index.ts)
- Drops the inlined `renderPluginUI` + `createNode` + `makeDraggable`
  + `ALLOWED_ELEMENT_TYPES` block (170 lines).
- Imports `renderPluginUI` from `@tlsn/host-extension/content`.
- The RENDER_PLUGIN_UI message handler passes a tiny
  `onPluginAction` closure that fires `browser.runtime.sendMessage`
  with `{type: 'PLUGIN_UI_CLICK', onclick, windowId}` — the existing
  contract the background expects, just lifted out of the renderer
  itself.

Content/index.ts shrinks from 346 → 176 lines. Still contains:
- the page-script injection (content.bundle.js → page scope)
- the background → content message routing (OFFSCREEN_LOG /
  PROVE_PROGRESS / RENDER_PLUGIN_UI / GET_PAGE_INFO / window.tlsn
  bridge from page)

Verification
- host-extension: 150/150 tests pass
- extension: 34/34 tests pass

The architecture-level extraction is now essentially done. Remaining:
extend the build-tlsn-host skill with an extension flow (the
packages/extension itself is already the reference consumer).
