+++ packages/host-extension/src/content/index.ts
@@ -1,0 +1,9 @@
+export {
+  renderPluginUI,
+  unmountPluginUI,
+  createNode,
+  makeDraggable,
+  ALLOWED_ELEMENT_TYPES,
+  type PluginActionHandler,
+  type RenderPluginUIOptions,
+} from './PluginRenderer.js';
